Adding the tests from the hydra repo to this repo
