+++ src/ResourceManagement/Commerce/UsageAggregates.Tests/UsageAggregates.Tests.csproj
@@ -17,7 +17,7 @@
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RecordedDelegatingHandler.cs" />
-    <Compile Include="ScenarioTests\GetUsageForSubscription.cs" />
+    <Compile Include="ScenarioTests\UsageAggregatesScenarioTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
